Extract cloud conflict review helpers
